LightsV1
Information: 

This circuit can be used to turn anything on and off, as long as it does not exceed the power threshold of the relay (which it shouldn’t).

I used a MOSFET and Relay combo. 
	A MOSFET cannot be used to control a AC current. A relay can.
	The MOSFET stays on when a positive voltage is applied, through Gate. When the MOSFET is on, the relay turns on. When the relay is turned on, the 120V AC can flow.
	The Flyback Diode is to prevent the EMF current from affecting the rest of the circuit. The relay uses an electromagnet.
	
To construct this circuit, you must cut the lamp cord (get an extension cord). The relay will be spliced in.
	Make sure to wire the relay into the positive wire. This is often the white wire.

 GPIO pin should not really matter.